Move navbar, switcher and useDarkSide to components folder
