+++ src/layout/MainLayout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Navbar from '../Navbar';
+import Navbar from '../components/Navbar';
 
 export default function MainLayout({ children }) {
   return (
